Move TemplatePreview to a separate file.
